Reestructura boilerplate
