+++ src/appFunctions/fbConfig.js
@@ -1,13 +1,60 @@
 import {getFirestore, addDoc, collection, getDocs, query, where, deleteDoc, doc, setDoc, getDoc, updateDoc} from '../lib/firestoreConfig.js'; //eslint-disable-line
+import {initializeApp, firebaseConfig, getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, onAuthStateChanged, signOut, signInWithPopup, GoogleAuthProvider} from '../lib/firebaseConfig.js'; //eslint-disable-line
 import { onNavigate } from '../routes/main.js'; //eslint-disable-line
-import { currentUsermail, currentUserid } from './firebase.js'; //eslint-disable-line
 
+initializeApp(firebaseConfig);
 const db = getFirestore();
+const auth = getAuth();
 let docId = '';
+let currentUserid = '';
+export let currentUsermail = ''; //eslint-disable-line
 let currentName = ''; //eslint-disable-line
-export let errorMessage = '';//eslint-disable-line
+export let errorMessage = "";//eslint-disable-line
 const timeElapsed = Date.now();
 const today = new Date(timeElapsed);
+
+export const createUser = (email, password) => {
+  createUserWithEmailAndPassword(auth, email, password)
+    .then((userCredential) => {
+      const user = userCredential.user;//eslint-disable-line
+    })
+    .catch((error) => {
+      const errorCode = error.code;
+      errorMessage = error.message;
+      if (errorCode === 'auth/invalid-email') {
+        alert('email invalido');//eslint-disable-line
+      } else if (errorCode === 'auth/weak-password') {
+        alert('contraseña invalida');//eslint-disable-line
+      } else if (errorCode === 'auth/missing-email') {
+        alert('falta correo');//eslint-disable-line
+      } else if (errorCode === 'auth/internal-error') {
+        alert('falta correo');//eslint-disable-line
+      }
+    });
+};
+export function authenticUser() {
+  onAuthStateChanged(auth, (user) => {
+    if (user) {
+      currentUserid = user.uid;
+      currentUsermail = user.email;
+    } else if (user === null) {
+      onNavigate('/');
+    }
+  });
+} authenticUser();
+
+export const loginUser = (email, password) => {
+  signInWithEmailAndPassword(auth, email, password)
+    .then((userCredential) => {//eslint-disable-line
+      // console.log(userCredential);
+      onNavigate('/post');
+    })
+    .catch((error) => {
+      const errorCode = error.code;//eslint-disable-line
+      errorMessage = error.message;//eslint-disable-line
+      alert(errorMessage);
+    });
+};
 
 export async function getName() {//eslint-disable-line
   try {
@@ -18,6 +65,14 @@ export async function getName() {//eslint-disable-line
     console.error('Error adding document: ', e);//eslint-disable-line
   }
 }
+getName();
+
+export const logoutUser = () => {
+  signOut(auth).then(() => {
+  }).catch((error) => {//eslint-disable-line
+  // An error happened.
+  });
+};
 
 export async function createProfile(name, lastName, email) {
   try {
@@ -28,6 +83,7 @@ export async function createProfile(name, lastName, email) {
     });
     docId = docRef.id;
     console.log('Document written with ID: ', docId); //eslint-disable-line
+    onNavigate('/login');
   } catch (e) {
     console.error('Error adding document: ', e);//eslint-disable-line
     console.log(console.error);//eslint-disable-line
@@ -71,6 +127,21 @@ export async function editPost(id, editedPost, editedPlace, editedHours, editedM
     Money: editedMoney,
   });
 }
+
+export const loginGoogle = () => {
+  const provider = new GoogleAuthProvider();
+  signInWithPopup(auth, provider)
+    .then((result) => {
+      const credential = GoogleAuthProvider.credentialFromResult(result);//eslint-disable-line
+      onNavigate('/post');
+    }).catch((error) => {
+      const errorCode = error.code;//eslint-disable-line
+      const errorMessage = error.message;//eslint-disable-line
+      const email = error.email;//eslint-disable-line
+      // The AuthCredential type that was used.
+      const credential = GoogleAuthProvider.credentialFromError(error);//eslint-disable-line
+    });
+};
 
 export async function updateLike(id, arrayLikes) {
   const i = arrayLikes.indexOf(currentUsermail);
