+++ test/appFunctions/fbConfig.spec.js
